refactor: Mudando nome da main
